Average.py

+++ Average.py
@@ -2,9 +2,9 @@
 # Date: 4/6/2021
 # Description: Asks the user for five numbers and then prints out the average of those numbers.
 print("Enter Five Numbers.")
-num_1 = float(input())
-num_2 = float(input())
-num_3 = float(input())
-num_4 = float(input())
-num_5 = float(input())
+num_1 = int(input())
+num_2 = int(input())
+num_3 = int(input())
+num_4 = int(input())
+num_5 = int(input())
 print("The average of those numbers is:", (num_1 + num_2 + num_3 + num_4 + num_5) / 5)
